crawler/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
paho-mqtt==2.1.0
pubsub==0.1.2
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
Pypubsub==4.0.3
rdflib==7.1.3
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SPARQLWrapper==2.0.0
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
